feat: Imp Satellite Detail Screen
Refs: SA-6

+++ domain/src/main/java/com/enciyo/domain/model/Positions.kt
@@ -10,6 +10,6 @@ data class Positions(
 )
 
 data class Position(
-    val posX: Long,
-    val posY: Long,
+    val posX: Double,
+    val posY: Double,
 )
